+++ src/OrchardCore/OrchardCore.Abstractions/Modules/IModularTenantEvents.cs
@@ -10,28 +10,3 @@ public interface IModularTenantEvents
     Task TerminatedAsync();
     Task RemovingAsync(ShellRemovingContext context);
 }
-
-public class ModularTenantEvents : IModularTenantEvents
-{
-    public virtual Task ActivatedAsync()
-    {
-        return Task.CompletedTask;
-    }
-
-    public virtual Task ActivatingAsync()
-    {
-        return Task.CompletedTask;
-    }
-
-    public virtual Task TerminatedAsync()
-    {
-        return Task.CompletedTask;
-    }
-
-    public virtual Task TerminatingAsync()
-    {
-        return Task.CompletedTask;
-    }
-
-    public virtual Task RemovingAsync(ShellRemovingContext context) => Task.CompletedTask;
-}
