+++ src/OrchardCore/OrchardCore.Abstractions/Modules/ModularTenantEvents.cs
@@ -1,0 +1,28 @@
+using OrchardCore.Environment.Shell.Removing;
+
+namespace OrchardCore.Modules;
+
+public class ModularTenantEvents : IModularTenantEvents
+{
+    public virtual Task ActivatedAsync()
+    {
+        return Task.CompletedTask;
+    }
+
+    public virtual Task ActivatingAsync()
+    {
+        return Task.CompletedTask;
+    }
+
+    public virtual Task TerminatedAsync()
+    {
+        return Task.CompletedTask;
+    }
+
+    public virtual Task TerminatingAsync()
+    {
+        return Task.CompletedTask;
+    }
+
+    public virtual Task RemovingAsync(ShellRemovingContext context) => Task.CompletedTask;
+}
