+++ test/OrchardCore.Tests/Apis/ContentManagement/DeploymentPlans/ContentStepLuceneQueryTests.cs
@@ -36,8 +36,8 @@ public class ContentStepLuceneQueryTests
 
         // Indexing of the content item happens in the background and may not be immediate available,
         // so we wait a bit.
-        // await context.WaitForOutstandingDeferredTasksAsync(TestContext.Current.CancellationToken);
-        await Task.Delay(10000, TestContext.Current.CancellationToken);
+        await context.WaitForOutstandingDeferredTasksAsync(TestContext.Current.CancellationToken);
+
         // Test
         var result = await context
             .GraphQLClient
