+++ test/OrchardCore.Tests/Apis/GraphQL/Queries/RecentBlogPostsQueryTests.cs
@@ -28,7 +28,7 @@ public class RecentBlogPostsQueryTests
 
         // Indexing of the content item happens in the background and may not be immediate available,
         // so we wait until the indexing is done before querying.
-        await Task.Delay(10000, TestContext.Current.CancellationToken);
+        await context.WaitForOutstandingDeferredTasksAsync(TestContext.Current.CancellationToken);
 
         var result = await context
             .GraphQLClient
